Use uploader extractor from utils
Greatly simplifies code
