+++ pyproject.toml
@@ -21,7 +21,7 @@ line_length=120                # corresponds to -w  flag
 multi_line_output=3            # corresponds to -m  flag
 include_trailing_comma=true    # corresponds to -tc flag
 skip_glob = '^((?!py$).)*$'    # this makes sort all Python files
-known_third_party = ["dacite", "more_itertools", "paho"]
+known_third_party = ["dacite", "paho"]
 
 [tool.poetry.dependencies]
 python = ">=3.7,<3.11"
